requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.2.1
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
executing==1.2.0
fastapi==0.112.0
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kubernetes==30.1.0
langchain==0.2.12
langchain-chroma==0.1.2
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.98
markdown-it-py==3.0.0
marshmallow==3.21.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.5.6
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.40.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
parso==0.8.3
pickleshare==0.7.5
pip==24.0
platformdirs==3.8.1
posthog==3.5.0
prompt-toolkit==3.0.39
protobuf==4.25.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
PyMySQL==1.1.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.0
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sse-starlette==2.1.3
stack-data==0.6.2
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.3.2
tqdm==4.66.5
traitlets==5.9.0
typer==0.12.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.2
uvicorn==0.30.5
watchfiles==0.23.0
wcwidth==0.2.6
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
